Project file is modified only when there are changes
At least for the latest XCode (11.5) XUnique seems to trigger a reload of the project file any time it runs, breaking code completion and bringing my computer to its knees every time I build as it then has to rebuild.

This seems to be because currently the file is changed in place and then reverted back to the old one if there are no changes. This intermediate change seems to get XCode to think the project changed even though XUnique reverts the change.

My changes make the script write the new project file to a separate file first and then only copies over to the project file if there are changes.